20230420 Maintenance
Gives some slight tweaks to the project by updating the README for
instructions to utilize a virtual environment, adding a
requirements.txt file (which was not previously available), as well as
creating the ability to utilize poetry (with an added poetry script
specification).

This does change `main.py` to `stixops.py`.