Fix poetry run script
Fixes the poetry script definition for initializing the application.
This changes the STIXOps.main to STIXOps.init_options for clarity.
Additionally added the `if __name__ == "__main__"` item with a core
main() function so that the script will run the same way if called by
python alone, instead of by poetry.